# refactor: Optimize compareEntries function for better performance