This PR optimizes the compareEntries function to improve performance while maintaining existing functionality and test compatibility.

## Summary

- Added early size comparison to avoid unnecessary iterations
- Optimized Map instance handling to prevent redundant Map creation
- Implemented WeakMap caching for complex key comparisons
- Maintained order-independence for object/map comparisons

## Check List

- [x] `pnpm run prettier` for formatting code and docs