@dai-shi 2024-12-17
https://github.com/pmndrs/zustand/actions/runs/12369221038/job/34521766886?pr=2908

**Size Change:** +227 B (+1.9%) 
**Total Size:** 12.2 kB
| Filename | Size | Change |  |
| :--- | :---: | :---: | :---: |
| `./dist/esm/vanilla/shallow.mjs` | 636 B | +113 B (+21.61%) | 🚨 |
| `./dist/vanilla/shallow.js` | 649 B | +114 B (+21.31%) | 🚨 |
<details><summary>ℹ️ <strong>View Unchanged</strong></summary>
| Filename | Size |
| :--- | :---: |
| `./dist/esm/index.mjs` | 64 B |
| `./dist/esm/middleware.mjs` | 3.58 kB |
| `./dist/esm/middleware/immer.mjs` | 210 B |
| `./dist/esm/react.mjs` | 320 B |
| `./dist/esm/react/shallow.mjs` | 193 B |
| `./dist/esm/shallow.mjs` | 82 B |
| `./dist/esm/traditional.mjs` | 411 B |
| `./dist/esm/vanilla.mjs` | 403 B |
| `./dist/index.js` | 244 B |
| `./dist/middleware.js` | 3.59 kB |
| `./dist/middleware/immer.js` | 229 B |
| `./dist/react.js` | [336](https://github.com/pmndrs/zustand/actions/runs/12369221038/job/34521766886?pr=2908#step:5:346) B |
| `./dist/react/shallow.js` | 208 B |
| `./dist/shallow.js` | 189 B |
| `./dist/traditional.js` | 423 B |
| `./dist/vanilla.js` | 416 B |
</details>
<a href="https://github.com/preactjs/compressed-size-action"><sub>compressed-size-action</sub></a>