@dai-shi 2024-12-17
Thanks for the suggestion. However, I don't think it's worth the complexity. We don't use `WeakMap` in our code base either.

If anyone is interested in experimenting more, consider creating a separate package. I can invite to https://github.com/zustandjs if they are interested.